refactor: #12 Cleanup convert codes
