+++ src/main.rs
@@ -5,6 +5,7 @@ use axum::{
 
 mod controllers;
 mod models;
+mod utils;
 
 #[tokio::main(flavor = "multi_thread", worker_threads = 16)]
 async fn main() {
